re-ordering some specs
