+++ test/e2e/data_setup/config/parallel-configs/all-features.dev.json
@@ -1,6 +1,7 @@
 {
     "setup": {
         "schemaConfigurations": [
+            "record/config.json",
             "record/copy-btn.config.json",
             "multi-permissions/config.json",
             "navbar/config.json"
