+++ test/e2e/data_setup/config/record/copy-btn.dev.json
@@ -1,6 +1,7 @@
 {
     "setup": {
         "schemaConfigurations": [
+            "record/config.json",
             "record/copy-btn.config.json"
         ],
         "schema": "editable-id"
